task: Move entity types to a new file types
