+++ src/moviesFreak/entities/index.ts
@@ -1,4 +1,4 @@
 export { default as Movie } from './movie';
 export { default as Session } from './session';
 export { default as User, UserPassword } from './user';
-export { default as WatchHub, WatchHubPrivacy } from './watchHub';
+export { default as WatchHub } from './watchHub';
