+++ src/moviesFreak/entities/types.ts
@@ -1,0 +1,5 @@
+export enum WatchHubPrivacy {
+  PUBLIC = 'public',
+  PRIVATE = 'private',
+  SHARED = 'shared'
+}
